self application of databinding concepts
during the course I was prompted to create this ng-project that adhered to the following requirements. It helped me to solidify my understanding of databinding and inter component communication. I may need to refer back to it for better understanding of local template variables. this was an interesting practical demonstration of structural dom manipulation through the use of *ngFor and *ngIf I had thought the solution would have called for more thorough logic but as i continued experimenting it turned out the use of *ngIf would be enough to hide the counter odd-even component in a cycle.

  <li>Create three new components: GameControl, Odd and Even</li>
        <li>The GameControl Component should have buttons to start and stop the game</li>
        <li>When starting the game, an event (holding a incrementing number) should get emitted each second (ref = setInterval())</li>
        <li>The event should be listenable from outside the component</li>
        <li>When stopping the game, no more events should get emitted (clearInterval(ref))</li>
        <li>A new Odd component should get created for
          every odd number emitted, the same should happen for the Even
          Component (on even numbers)</li>
        <li>Simply output Odd - NUMBER or Even - NUMBER in the two components</li>
        <li>Style the element (e.g. paragraph) holding your output text differently in both components</li>